- add action to set multiple variables - add actions to set cursor position - add eval to levelObject's getProperty for Players, Items, SimpleLevelObject - add level.enableHover - add new random number generator (with normal distribution) - add prefix/suffix price classifiers to ItemClass - add relativeCoords to drawable.setPosition action - add spell level to spells - fix saving empty inventory removing it from player - fix trim functions - Formula refactor to support dynamic formulas - increase item properties size - refactor item price - rework useItem to use random in formulas instead of having value and ValueMax - save player's selected spell - update canUseObject to use formulas for items too - update Formula class - update gamefiles - update gamefiles (cathedral map) - update gamefiles (item prefix/suffic prices) - update gamefiles (spell book UI) - update gamefiles (spell UI) - update GameUtils::replace* functions - update loadJson to load from file and use parameters - update Spells - update TextUtils - update variable.set* actions

+++ src/SFML/SFMLUtils.cpp
@@ -75,7 +75,7 @@ namespace SFMLUtils
 		// A new view (with a new viewport set) is returned.
 
 		float windowRatio = (float)windowSize.x / (float)windowSize.y;
-		float viewRatio = view.getSize().x / (float)view.getSize().y;
+		float viewRatio = view.getSize().x / view.getSize().y;
 
 		bool horizontalSpacing = true;
 		if (windowRatio < viewRatio)
